// src/llamacpp-client/types.gen.ts
export type _Error = {
  error?: {
    code?: number;
    message?: string;
    type?: string;
  };
};